move caching methods to an individual trait
